+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased](https://github.com/Instabug/Instabug-Flutter/compare/v14.0.0...dev)
+
+### Added
+
+- Add support enable/disable stop capturing network body. ([#561](https://github.com/Instabug/Instabug-Flutter/pull/561))
+
 ## [14.1.0](https://github.com/Instabug/Instabug-Flutter/compare/v14.0.0...v14.1.0) (January 2, 2025)
 
 ### Changed