+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Add support enable/disable stop capturing network body. ([#561](https://github.com/Instabug/Instabug-Flutter/pull/561))
+- Add support enable/disable capturing network body. ([#561](https://github.com/Instabug/Instabug-Flutter/pull/561))
 
 ## [14.1.0](https://github.com/Instabug/Instabug-Flutter/compare/v14.0.0...v14.1.0) (January 2, 2025)