+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [Unreleased](https://github.com/Instabug/Instabug-Flutter/compare/v14.1.0...dev)
+
+### Fixed
+
+- Fixed an issue with `SetReproStepsConfig` on Android platform ([#543](https://github.com/Instabug/Instabug-Flutter/pull/543)).
 
 ## [14.1.0](https://github.com/Instabug/Instabug-Flutter/compare/v14.0.0...v14.1.0) (January 2, 2025)