+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## [Unreleased](https://github.com/Instabug/Instabug-React-Native/compare/v14.1.0...dev)
+## [Unreleased](https://github.com/Instabug/Instabug-Flutter/compare/v14.1.0...dev)
 
 ### Changed